set maxFileSize to NSUInteger; code style changes;

+++ QMShareExtension/Common classes/QMAttachmentProvider.h
@@ -16,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface QMAttachmentProviderSettings : NSObject
 
-@property (assign, nonatomic) CGFloat maxFileSize;  //In megabytes
+@property (assign, nonatomic) NSUInteger maxFileSize;  //In megabytes
 @property (assign, nonatomic) CGFloat maxImageSize; //In pixels
 @property (assign, nonatomic) CGFloat imageQuality; // from 0 to 1.0
 @end
